chore: refactor tab tokens
